+++ packages/client/src/App.tsx
@@ -6,7 +6,6 @@ import {
 } from 'react-router-dom'
 import Header from '@/components/header/Header'
 import Main from '@/pages/main/main'
-import Game from '@/pages/game/game'
 import Forum from '@/pages/forum/forum'
 import LeaderBord from '@/pages/leader-board/leader-board'
 import Error404 from '@/pages/page-error/Error404'
@@ -16,6 +15,7 @@ import Documentation from '@/pages/documentation/documentation'
 import { Profile } from '@/pages/profile/profile'
 import { ROUTES } from './routes'
 import style from './styles/index.module.scss'
+import { Game } from './pages/game/game'
 
 function App() {
   // TODO: приватные роуты будут готовы после того, как будет готова авторизация
