+++ packages/client/src/canvas/index.module.scss
@@ -1,0 +1,5 @@
+.pixelRatio {
+  width: 1350px;
+  height: 800px;
+  background: black;
+}
