+++ packages/client/src/pages/game/components/index.module.scss
