+++ packages/client/src/pages/game/components/main.tsx
