--- packages/client/src/pages/game/components/stanbdy/index.module.scss
@@ -1,73 +1,0 @@
-@use '@/styles/mixins';
-@use '@/styles/animationMixins';
-
-$text-size: 6rem;
-$wait-color: rgb(39 44 60 / 87%);
-$win-color: rgb(52 0 162 / 65%);
-$loser-color: rgb(162 0 0 / 65%);
-
-.statdby {
-  position: absolute;
-
-  display: flex;
-  inset: 0;
-
-  &Initialize {
-    background-color: var(--color-bg-body);
-  }
-
-  &Pause {
-    background-color: $wait-color;
-  }
-
-  &FinishWinner {
-    background-color: $win-color;
-  }
-
-  &FinishLoser {
-    background-color: $loser-color;
-  }
-
-  &Body {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: auto;
-  }
-
-  &Text {
-    font-size: $text-size;
-  }
-
-  @include animationMixins.animation-wrapper {
-    &Initialize &Text {
-      animation: 5s infinite alternate slidein;
-    }
-  }
-
-  &ButtonCont {
-    margin-top: 1rem;
-  }
-
-  &ButtonCont > * {
-    min-width: 6rem;
-  }
-
-  &ButtonCont > * ~ * {
-    margin-left: 1rem;
-  }
-}
-
-@keyframes slidein {
-  0% {
-    color: var(--color-gradient-1);
-  }
-
-  20% {
-    color: var(--color-gradient-2);
-  }
-
-  100% {
-    color: var(--color-gradient-3);
-  }
-}
