+++ packages/client/src/pages/game/index.module.scss
@@ -1,16 +1,5 @@
-@use '@/styles/mixins';
-
 .game {
   display: flex;
-
-  flex-grow: 1;
-
-  &Body {
-    position: relative;
-
-    width: 1100px;
-    height: 600px;
-
-    margin: auto;
-  }
+  align-items: center;
+  flex-direction: column;
 }
